Proposal: Require cross-organizational seconds for reviewer nominations https://bugs.webkit.org/show_bug.cgi?id=47597
Reviewed by Darin Adler.

This changes the committer/reviewer policy to require an additional
reviewer supporting a reviewer nomination to work for a different
company and different project affiliation from the nominee.

This means reviewer nominations would require four supporting reviewers,
one of which knows the reviewer through the community (irc, email, bugs)
ensuring that the nominee has the open-source collaboration skills
we so highly value here in WebKit.

* coding/commit-review-policy.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@76367 268f45cc-cd09-0410-ab3c-d52691b4dbfc