Double-check that modifying the registry worked
On Windows Vista/7 with UAC enabled, regtool will fail to modify the registry, but it will
still return a successful exit code. So we double-check here that the value we tried to
write to the registry was really written.

Fixes <http://webkit.org/b/55227> old-run-webkit-tests falsely claims to be saving crash
logs on Windows Vista/7

Reviewed by Darin Adler.

* Scripts/old-run-webkit-tests:
(writeRegistryString): After trying to write the string value, check that it was really
written.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@79698 268f45cc-cd09-0410-ab3c-d52691b4dbfc