<rdar://problem/9005982> Flash of white when a WKView with composited content moves on-screen https://bugs.webkit.org/show_bug.cgi?id=57522
Reviewed by Maciej Stachowiak.

* UIProcess/DrawingAreaProxyImpl.cpp:
(WebKit::DrawingAreaProxyImpl::exitAcceleratedCompositingMode): Invalidate the current
backing store state, so that we get an update as soon as the page enters compositing mode
next.
* WebProcess/WebPage/DrawingAreaImpl.cpp:
(WebKit::DrawingAreaImpl::sendDidUpdateBackingStoreState): Added a call to
LayerTreeHost::forceRepaint() in order to flush and synchronize the layers
before sending the update message to the UI process.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@82543 268f45cc-cd09-0410-ab3c-d52691b4dbfc