https://bugs.webkit.org/show_bug.cgi?id=58151 Crash after closing a Webview, in WebKit!WebCore::WindowMessageBroadcaster::SubclassedWndProc
Reviewed by Adam Roben.

* UIProcess/win/WebView.cpp:
(WebKit::WebView::close):
Resolve a difference between old WebKit WebView and WebKit2 WebView by destroying and clearing
out WebKit2 WebView's HWND when closing. Not doing so was causing the WebView to remain in the
list of WindowMessageBroadcaster listeners, and after closing the WebView, WindowMessageBroadcaster
would message a dead WebView and crash.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@83357 268f45cc-cd09-0410-ab3c-d52691b4dbfc