CSSSelector double frees https://bugs.webkit.org/show_bug.cgi?id=56124 <rdar://problem/9119036>
Reviewed by Simon Fraser.

In all stacks this is occuring when the document is kept alive by the guard ref only
and the node holding it is deleted.

As a speculative fix, clear the document stylesheets when it turns into a guard ref zombie.
This has a pretty good chance of fixing the actual bug. At minimum it will move the
crash to an earlier, more easily traceble point.

No tests, crash stacks only, no idea how to reproduce.

* dom/Document.cpp:
(WebCore::Document::removedLastRef):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@84151 268f45cc-cd09-0410-ab3c-d52691b4dbfc