Text does not split on white space when typing to the edge of window. https://bugs.webkit.org/show_bug.cgi?id=59968 <rdar://problem/9112683>
Reviewed by Dan Bernstein.

Source/WebCore: 

Test: editing/inserting/typing-at-end-of-line.html

This is a regression introduced when we converted the line box tree to
floating point in r78846. In findNextLineBreak, there was still one place where
the character width was treated as int and truncated and we were inconsistent
in the way we treated a character that did not fit entirely in the line.

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::findNextLineBreak):

LayoutTests: 

* editing/inserting/typing-at-end-of-line.html: Added.
* platform/mac/editing/inserting/typing-at-end-of-line-expected.txt: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@85535 268f45cc-cd09-0410-ab3c-d52691b4dbfc