[Qt] Layout tests may timeout if loadFinished receives false https://bugs.webkit.org/show_bug.cgi?id=60064
Reviewed by NOBODY Kenneth Rohde Christiansen.

When using layoutTestController.queueLoad(), if QWebPage::loadFinished(bool) receives false
then the layout test may timout. It happens that QWebPage::loadFinished() receives false when
a response is ignored, like it should be for 204 responses. It is a problem for creating
tests for 204 responses for example.

This change has been tested and it causes no regressions in layout tests.

* DumpRenderTree/qt/LayoutTestControllerQt.cpp:
(LayoutTestController::maybeDump):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@85667 268f45cc-cd09-0410-ab3c-d52691b4dbfc