A WKView with plug-ins paints white when first moved into a window https://bugs.webkit.org/show_bug.cgi?id=60511
Reviewed by Darin Adler.

The UI process’s wait for DidUpdateBackingStoreState would always time out because the Web process
was blocked on a synchronous message the UI process trying to get the connections for plug-ins in
the view. This is avoided by deferring the call to Page::setCanStartMedia().

* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::WebPage): Initialize the new m_setCanStartMediaTimer.
(WebKit::WebPage::setCanStartMediaTimerFired): Added. Calls setCanStartMedia(true) when the timer
fires.
(WebKit::WebPage::setIsInWindow): When going out of a window, stop the timer. When going into a
window, start it.
* WebProcess/WebPage/WebPage.h:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@86103 268f45cc-cd09-0410-ab3c-d52691b4dbfc