Fix for <rdar://problem/9253454> Solar Walk For Mac: Info window is blank (HTML5 parser) https://bugs.webkit.org/show_bug.cgi?id=60685
Reviewed by Darin Adler and Adam Barth.

Source/WebCore: 

Detect the Solar Walk application so we can add targeted parser quirks.

* WebCore.exp.in:
* platform/RuntimeApplicationChecks.cpp: (WebCore::applicationIsSolarWalk):
* platform/RuntimeApplicationChecks.h:

Source/WebKit: 

* WebKit.xcodeproj/project.pbxproj: Added Misc/SolarWalkQuirksUserScript.js and copy it to Resources during the build phase. 

Source/WebKit/mac: 

Solar Walk uses a self closing title tag, so to match the behavior of the old parser, 
we use a UserScript to write the contents of the title element to the end of the document so it can be re-parsed.

* Misc/SolarWalkQuirksUserScript.js: Added.
* WebView/WebView.mm:
(needsSolarWalkQuirksScript):
(leakSolarWalkQuirksUserScriptContents):
(-[WebView _injectSolarWalkQuirksScript]):
(-[WebView _commonInitializationWithFrameName:groupName:usesDocumentViews:]):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@86407 268f45cc-cd09-0410-ab3c-d52691b4dbfc