Frame::visiblePositionForPoint returns null position if the point is over an imagemap. https://bugs.webkit.org/show_bug.cgi?id=62562 <rdar://problem/9313885>
Patch by Morgan Winer <mwiner@apple.com> on 2011-06-13
Reviewed by Enrica Casucci.

The innerNode() of an imageMap is an area, and doesn't have a renderer and
therefore can't report a position.
Every result's innerNonSharedNode() should be the same as the innerNode(),
except for imageMaps, which will have the actual image as its innerNonSharedNode(),
and will therefore have a renderer and can report a position.

There are no layout tests for this, since it can be tested only at API level.

* page/Frame.cpp:
(WebCore::Frame::visiblePositionForPoint): Changing innerNode() to innerNonSharedNode().



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@88718 268f45cc-cd09-0410-ab3c-d52691b4dbfc