Web process always pauses on launch https://bugs.webkit.org/show_bug.cgi?id=63487
Reviewed by Oliver Hunt.

In the absence of the WEBKIT2_PAUSE_WEB_PROCESS_ON_LAUNCH environment variable, only pause the web process on
launch if the shift, control, and alt keys are all held down. This regressed in r89865, which always caused the
web process to pause if the shift and alt keys were not held down.

* WebProcess/WebKitMain.cpp:
(pauseProcessIfNeeded): Add some parentheses.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@89865 268f45cc-cd09-0410-ab3c-d52691b4dbfc