https://bugs.webkit.org/show_bug.cgi?id=63947 ASSERT running Object.preventExtensions(Math.sin)
Reviewed by Oliver Hunt.

This is due to calling scope() on a hostFunction as a part of
calling createPrototypeProperty to reify the prototype property.
But host functions don't have a prototype property anyway!

Source/JavaScriptCore: 

Prevent callling createPrototypeProperty on a host function.

* runtime/JSFunction.cpp:
(JSC::JSFunction::createPrototypeProperty):
(JSC::JSFunction::preventExtensions):

LayoutTests: 

Add test case for calling preventExtensions on a host function.

* fast/js/preventExtensions-expected.txt:
* fast/js/script-tests/preventExtensions.js:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@90402 268f45cc-cd09-0410-ab3c-d52691b4dbfc