Upstream QtWebKit/N9 changes related to visibleContentRect https://bugs.webkit.org/show_bug.cgi?id=64589
Reviewed by Simon Hausmann.

Source/WebCore:

Rename setActualVisibleContentRect to setFixedVisibleContentRect
as that makes it more obvious that it is an override. It is also
consistent with the setFixedLayout which is often used in
conjunction with it.

Make visibleContentRect return the fixed value if set, and remove
all calls to actualVisibleContentRect.

Also updated the documentation.

This is similar to what we have on our QtWebKit/N9 branch and is
pretty well tested.

* platform/ScrollView.cpp:
(WebCore::ScrollView::visibleContentRect):
If a fixed visibleContentRect is set, return that.
Remove the code returning the contents size when in
paintsEntireContents mode as that is wrong and already
ifdeffed out for EFL.
(WebCore::ScrollView::setScrollPosition):
Do not try to be smart and update the fixed visibleContentsRect
automatically, as the embedder might ignore the scroll request
for various reasons.
* platform/ScrollView.h:
(WebCore::ScrollView::setFixedVisibleContentRect):
(WebCore::ScrollView::fixedVisibleContentRect):

Source/WebKit/qt:

* Api/qwebpage.cpp:
(QWebPage::setActualVisibleContentRect):
Change to use setFixedVisibleContentRect.
* WebCoreSupport/FrameLoaderClientQt.cpp:
(WebCore::FrameLoaderClientQt::transitionToCommittedForNewPage):
Only set fixedVisibleContentRect for the mainframe.
* symbian/eabi/QtWebKitu.def:

Source/WebKit2:

Rename setActualVisibleContentRect to setFixedVisibleContentRect
as that makes it more obvious that it is an override. It is also
consistent with the setFixedLayout which is often used in
conjunction with it.

* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::setFixedVisibleContentRect):
* UIProcess/WebPageProxy.h:
* UIProcess/qt/qtouchwebpageproxy.cpp:
(QTouchWebPageProxy::setVisibleArea):
* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::transitionToCommittedForNewPage):
Only set fixedVisibleContentRect for the mainframe.
* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::setFixedVisibleContentRect):
* WebProcess/WebPage/WebPage.h:
* WebProcess/WebPage/WebPage.messages.in:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@91064 268f45cc-cd09-0410-ab3c-d52691b4dbfc