RenderText::absoluteRectsForRange() and absoluteQuadsForRange() have nearly duplicate code https://bugs.webkit.org/show_bug.cgi?id=62478
Reviewed by Simon Fraser.

Source/WebCore:

Test: fast/dom/Range/getClientRects.html

* platform/graphics/FloatRect.h:
(WebCore::FloatRect::isZero):
Add izZero method, unlike isEmpty this checks if both the width and the
height are zero.

* rendering/RenderText.cpp:
(WebCore::absoluteQuadForTextBox):
(WebCore::RenderText::absoluteRectsForRange):
(WebCore::RenderText::absoluteQuadsForRange):
Break duplicate code out of absoluteRectsForRange and
absoluteQuadsForRange into shared static function.

Fix what I presume to be a bug in the absoluteQuadsForRange where the
logicalHeight was always used to set the size in absoluteQuadForTextBox
as oppsued to the absoluteRectsForRange implementation that uses the
logicalWidth or the logicalHeight depending on whether it's horizontal or
vertical.

LayoutTests:

* fast/dom/Range/getClientRects-expected.txt:
* fast/dom/Range/getClientRects.html:
Add test for vertical text.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@91763 268f45cc-cd09-0410-ab3c-d52691b4dbfc