A few purely stylistic modifications to visible_units.cpp https://bugs.webkit.org/show_bug.cgi?id=65723
Patch by Van Lam <vanlam@google.com> on 2011-08-04
Reviewed by Ryosuke Niwa.

Renamed greatestValueUnder to greatestOffsetUnder, positionIsInsideBox
to positionIsInBoxButNotOnBoundary (to avoid confusion with
positionIsInBox, which is just a getInlineBoxAndOffset check).
Removed use of invalidOffset as an error value in greatestOffsetUnder
and smallestOffsetAbove since semantically it should only be used to
check if it makes sense to compare offsets in a single box.

* editing/visible_units.cpp:
(WebCore::greatestOffsetUnder):
(WebCore::smallestOffsetAbove):
(WebCore::positionIsInBoxButNotOnBoundary):
(WebCore::leftWordPositionAcrossBoundary):
(WebCore::rightWordPositionAcrossBoundary):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92431 268f45cc-cd09-0410-ab3c-d52691b4dbfc