BaselineOptimizer created the wrong baseline for fast/js/regexp-overflow.html https://bugs.webkit.org/show_bug.cgi?id=65891
Reviewed by Eric Seidel.

The problem was that platform/chromium contained a bogus expectation
file that needed to be removed, but by the time we got around to
removing it, we'd already moved the correct baseline into its place.
After this patch, we copy the gold results into memory before
reshuffling things on disk (and we delete bad things before adding good
things).

* Scripts/webkitpy/common/checkout/baselineoptimizer.py:
* Scripts/webkitpy/common/checkout/baselineoptimizer_unittest.py:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@92669 268f45cc-cd09-0410-ab3c-d52691b4dbfc