Fix a bunch of minor bugs caught by the clang static analyzer in JavaScriptCore https://bugs.webkit.org/show_bug.cgi?id=66182
Reviewed by Dan Bernstein.

Fixes 10 warnings in JavaScriptCore and 2 in testapi.

* API/tests/testapi.c:
(main):
Remove dead variables.

* dfg/DFGGraph.cpp:
(JSC::DFG::Graph::dump):
Initialize hasPrinted and silence an unused warning by casting to void (Ok here
since it is debug code and I want to keep it clear that if other cases are added,
the hasPrinted flag would be needed).

* wtf/dtoa.cpp:
(WTF::d2b):
The variable "de" in the else block is always zero, so there is no reason to
use it.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@93017 268f45cc-cd09-0410-ab3c-d52691b4dbfc