WebProcess crashes during startup if libdispatch is initialized by WebProcessShim.dylib https://bugs.webkit.org/show_bug.cgi?id=66508 <rdar://problem/9828476>
Reviewed by Mark Rowe.

* mac/MainMac.cpp:
(closeUnusedFileDescriptors):
Check if a file descriptor is a kqueue and don't close it if that is the case. While this
isn't a complete fix, (it won't work if other initializers end up creating non-kqueue file descriptors)
it's good enough for Snow Leopard. For Lion, we should use the new posix_spawn API that lets you whitelist
file descriptors from the parent process.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@93420 268f45cc-cd09-0410-ab3c-d52691b4dbfc