Some compositing LayoutTests flakily crashing on Mac 10.6 https://bugs.webkit.org/show_bug.cgi?id=66740
Reviewed by Adam Barth.

Tools:

Null-check obj->pluginTest before dereferencing it. This can be null in chromium if the plugin load fails, for
example if the test attempts to load a PDF plugin the load will fail since we don't have such a thing in the
chromium configuration.

* DumpRenderTree/TestNetscapePlugIn/main.cpp:
(NPP_Destroy):

LayoutTests:

Remove = CRASH expectations for plugin tests that should no longer crash.

* platform/chromium/test_expectations.txt:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@93737 268f45cc-cd09-0410-ab3c-d52691b4dbfc