TypedArrays don't ensure that denormalised values are normalised https://bugs.webkit.org/show_bug.cgi?id=67178
Reviewed by Gavin Barraclough.

../../../../Volumes/Data/git/WebKit/OpenSource/LayoutTests:

Add test to ensure that we create a non-signalling nan when reading
a singaling nan from a typed array.

* fast/canvas/webgl/webgl-array-invalid-ranges-expected.txt:
* fast/canvas/webgl/webgl-array-invalid-ranges.html:

../../../../Volumes/Data/git/WebKit/OpenSource/Source/JavaScriptCore:

Add a couple of assertions to jsNumber() to ensure that
we block signaling NaNs

* runtime/JSValue.h:
(JSC::jsDoubleNumber):
(JSC::jsNumber):

../../../../Volumes/Data/git/WebKit/OpenSource/Source/WebCore:

Ensure that we convert singaling nans to silent nans when loading
from a typed array.

* bindings/scripts/CodeGeneratorJS.pm:
(GenerateImplementation):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@94095 268f45cc-cd09-0410-ab3c-d52691b4dbfc