[Qt] Crash when dragging google maps. https://bugs.webkit.org/show_bug.cgi?id=68223
Source/WebCore: 

The custom image for a cursor might not be ready when WebKit trying
to use it to construct the cursor in Qt porting. This patch is using
arrow image to substitute the unready custom image.

Reviewed by Adam Barth.

Test: fast/css/crash-on-custom-cursor-when-loading.html

* platform/qt/CursorQt.cpp:
(WebCore::createCustomCursor):
(WebCore::Cursor::ensurePlatformCursor):

LayoutTests: 

Test case to use an unready custom cursor.

Reviewed by Adam Barth.

* fast/css/crash-on-custom-cursor-when-loading-expected.txt: Added.
* fast/css/crash-on-custom-cursor-when-loading.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95275 268f45cc-cd09-0410-ab3c-d52691b4dbfc