Directory Upload: parent path can truncate first char of the correct path https://bugs.webkit.org/show_bug.cgi?id=66695
Source/WebCore:

In some cases, if the parent path we compute ends with a separator
character like '/' or 'C:\', by adding one in an attempt to grab the
subsequent separator we actually grab a character of the real path,
which is incorrect.

Reviewed by Darin Fisher.

* html/FileInputType.cpp:
(WebCore::FileInputType::setFileList):

LayoutTests:

Add a test case for when the parent path is '/'.

Reviewed by Darin Fisher.

* fast/forms/input-file-directory-upload.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@95582 268f45cc-cd09-0410-ab3c-d52691b4dbfc