feColorMatrix saturation is not limited to range 0..1 anymore https://bugs.webkit.org/show_bug.cgi?id=69245
Source/WebCore: 

Reviewed by Nikolas Zimmermann.
        
Opera doesn't limit the range of values for saturation on feColorMatrix from 0 to 1. The limitation
was also removed from the new Filter Effects 1.0 specification. Values outside this range lead to
under- or oversaturation of the filter input image.
https://dvcs.w3.org/hg/FXTF/raw-file/tip/filters/publish/Filters.html#feColorMatrixElement

Test: svg/filters/feColorMatrix-saturate.svg

* svg/SVGFEColorMatrixElement.cpp:
(WebCore::SVGFEColorMatrixElement::build):

LayoutTests: 

Reviewed by Nikolas Zimmermann.

Added new test case to test under- or oversaturation of the filter input image if the value
is not in the range of 0..1.
feColorMatrix-values needed a change, since it relies to the old specification text.

* platform/mac/svg/filters/feColorMatrix-values-expected.png:
* platform/mac/svg/filters/feColorMatrix-values-expected.txt:
* svg/filters/feColorMatrix-saturate-expected.png: Added.
* svg/filters/feColorMatrix-saturate-expected.txt: Added.
* svg/filters/feColorMatrix-saturate.svg: Added.
* svg/filters/feColorMatrix-values.svg:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@96498 268f45cc-cd09-0410-ab3c-d52691b4dbfc