WebSocket: Add extensions attribute https://bugs.webkit.org/show_bug.cgi?id=69408
Reviewed by Kent Tamura.

Source/WebCore:

Add WebSocket's "extensions" attribute as specified in the WebSocket API
<http://dev.w3.org/html5/websockets/#the-websocket-interface>.

WebSocket protocol extension is not implemented at all, thus this attribute
always returns an empty string.

Test: http/tests/websocket/tests/hybi/extensions.html
      http/tests/websocket/tests/hixie76/undefined-attributes.html (updated)

* websockets/WebSocket.cpp:
(WebCore::WebSocket::extensions):
If the hixie-76 protocol is chosen, we need to return "undefined" in order to
keep behavior of the ports which have not used the newer hybi protocol yet.
* websockets/WebSocket.h:
* websockets/WebSocket.idl:

LayoutTests:

* http/tests/websocket/tests/hixie76/undefined-attributes-expected.txt:
* http/tests/websocket/tests/hixie76/undefined-attributes.html:
Added "extensions" attribute to the list because it shouldn't be defined
when hixie-76 protocol is used.
* http/tests/websocket/tests/hybi/extensions-expected.txt: Added.
* http/tests/websocket/tests/hybi/extensions.html:
Added. Test whether .extensions attribute always returns an empty string.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@96785 268f45cc-cd09-0410-ab3c-d52691b4dbfc