When building with clang, enable -Wglobal-constructors and -Wexit-time-destructors https://bugs.webkit.org/show_bug.cgi?id=69586
Reviewed by Darin Adler.

../JavaScriptCore: 

* Configurations/Base.xcconfig:
Add -Wglobal-constructors and -Wexit-time-destructors when building with clang.

* JavaScriptCore.xcodeproj/project.pbxproj:
When building with clang, we don't need to run the check-for-global-initializers and
check-for-exit-time-destructors anymore.

* jsc.cpp:
(runInteractive):
Move interpreterName into runInteractive.

* wtf/StdLibExtras.h:
When building with clang, disable the -Wglobal-constructors and -Wexit-time-destructors
warnings around the variable declaration.

../JavaScriptGlue: 

* Configurations/Base.xcconfig:
Add -Wglobal-constructors and -Wexit-time-destructors when building with clang.
        
* JavaScriptGlue.xcodeproj/project.pbxproj:
When building with clang, we don't need to run the check-for-global-initializers and
check-for-exit-time-destructors anymore.

../WebCore: 

* Configurations/Base.xcconfig:
Add -Wglobal-constructors and -Wexit-time-destructors when building with clang.
        
* WebCore.xcodeproj/project.pbxproj:
When building with clang, we don't need to run the check-for-global-initializers and
check-for-exit-time-destructors anymore.
        
* platform/graphics/filters/FEGaussianBlur.cpp:
(gaussianKernelFactor):
(WebCore::FEGaussianBlur::calculateKernelSize):
(WebCore::FEGaussianBlur::calculateStdDeviation):
Change the gaussian kernel factor global to be an inline function instead.

../WebKit: 

* WebKit.xcodeproj/project.pbxproj:
When building with clang, we don't need to run the check-for-global-initializers and
check-for-exit-time-destructors anymore.

../WebKit/mac: 

* Configurations/Base.xcconfig:
Add -Wglobal-constructors and -Wexit-time-destructors when building with clang.

../WebKit2: 

* Configurations/Base.xcconfig:
Add -Wglobal-constructors and -Wexit-time-destructors when building with clang.

* WebKit2.xcodeproj/project.pbxproj:
When building with clang, we don't need to run the check-for-global-initializers and
check-for-exit-time-destructors anymore.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@96881 268f45cc-cd09-0410-ab3c-d52691b4dbfc