[GTK] Don't expose objects with unknown role on GTK https://bugs.webkit.org/show_bug.cgi?id=69561
Reviewed by Chris Fleizach.

Source/WebCore:

Ignore accessibility objects in accessibilityPlatformIncludesObject.

Test: platform/gtk/accessibility/unknown-roles-not-exposed.html

* accessibility/gtk/AccessibilityObjectAtk.cpp:
(WebCore::AccessibilityObject::accessibilityPlatformIncludesObject):
For unknown accessibility objects, return IgnoreObject.

LayoutTests:

New test to check that unknown objects are not exposed in GTK.

* platform/gtk/accessibility/unknown-roles-not-exposed-expected.txt: Added.
* platform/gtk/accessibility/unknown-roles-not-exposed.html: Added.

Added new expectation for new accessibility layout test introduced
in r96127 that was not possible to add before because of this bug.

* platform/gtk/accessibility/deleting-iframe-destroys-axcache-expected.txt: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@96922 268f45cc-cd09-0410-ab3c-d52691b4dbfc