[Chromium] Prevent fallback fonts from purging while measuring https://bugs.webkit.org/show_bug.cgi?id=69855
Reviewed by Eric Seidel.

When Chrome computes the page width, an international page
may bring in a fallback font. FontCache::getCachedFontData
expects the font cache to be locked, so the measure path
needs to include an instance of FontCachePurgePreventer.

Adding FontCachePurgePreventer suppresses a debug-time assert
when running PageCycler over international tests.

* src/WebFrameImpl.cpp:
(WebKit::WebFrameImpl::contentsPreferredWidth):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97257 268f45cc-cd09-0410-ab3c-d52691b4dbfc