https://bugs.webkit.org/show_bug.cgi?id=70064
The NPObjectWrapper class used by V8 bindings in Chromium to wrap the window
script object was allocating an instance of itself in the NPAllocate
implementation and returning this pointer. It should be returning a pointer
to the wrapped NPObject structure (NPProxyObject). The member function
getUnderlyingNPObject should return 0 if we fail to find the underlying NPObject
for the call. It was incorrectly returning a pointer to the same NPObject in
this case which could cause recursion. 

Patch by Anantanarayanan G Iyengar <ananta@chromium.org> on 2011-10-13
Reviewed by Nate Chapin.

No new tests as there is no change in functionality.

* bindings/v8/NPObjectWrapper.cpp:
(WebCore::NPObjectWrapper::getObjectForCall):
(WebCore::NPObjectWrapper::NPAllocate):
* bindings/v8/NPObjectWrapper.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97429 268f45cc-cd09-0410-ab3c-d52691b4dbfc