AX: webkit won't always send live region notifications https://bugs.webkit.org/show_bug.cgi?id=70030
Source/WebCore: 

Webkit purposefully does not send live region notifications if some AT client has not accessed the AX tree in between 
the last live region update (ostensibly to improve performance).
However, this is problematic if a screen reader misses one notification... it has no idea what happened and the only 
recourse is to rescan the page for changes. The correct thing to do is to always send notifications.

Reviewed by Darin Adler.

Test: platform/mac/accessibility/aria-liveregions-notifications-always-sent.html

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::childrenChanged):

LayoutTests: 

Reviewed by Darin Adler.

* platform/mac/accessibility/aria-liveregions-notifications-always-sent-expected.txt: Added.
* platform/mac/accessibility/aria-liveregions-notifications-always-sent.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97627 268f45cc-cd09-0410-ab3c-d52691b4dbfc