[Qt][WK2] ASSERT when displaying context menu in input field https://bugs.webkit.org/show_bug.cgi?id=69859
Patch by Dinu Jacob <dinu.jacob@nokia.com> on 2011-10-18
Reviewed by Simon Hausmann.

When creating a submenu in the context menu, the submenu of type
QWidget is statically cast to QObject before calling setParent on
it. QObject::setParent asserts if the object is a QWidget. So, call
setParent of QWidget itself.

* UIProcess/qt/WebContextMenuProxyQt.cpp:
(WebKit::WebContextMenuProxyQt::createContextMenu):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97760 268f45cc-cd09-0410-ab3c-d52691b4dbfc