Add getConstructData to the MethodTable https://bugs.webkit.org/show_bug.cgi?id=70163
Reviewed by Geoffrey Garen.

Source/JavaScriptCore:

Adding getConstructData to the MethodTable in order to be able to
remove all calls to getConstructDataVirtual soon.  Part of the process
of de-virtualizing JSCell.

* JavaScriptCore.exp:
* JavaScriptCore.vcproj/JavaScriptCore/JavaScriptCore.def:
* runtime/ClassInfo.h:

Source/WebCore:

No new tests.

Adding getConstructData to the MethodTable in order to be able to
remove all calls to getConstructDataVirtual soon.  Part of the process
of de-virtualizing JSCell.

* WebCore.exp.in:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97892 268f45cc-cd09-0410-ab3c-d52691b4dbfc