REGRESSION (r96823): Contextual menu closes immediately when control-clicking in Flash plug-in https://bugs.webkit.org/show_bug.cgi?id=70534 <rdar://problem/10308827>
Reviewed by Darin Adler.

Source/WebCore:

* plugins/PluginView.cpp: (WebCore::PluginView::handleEvent): Return true for contextmenu
event, so that plug-ins won't get a default WebKit context menu. We can't know if the
plug-in is handling mousedown (or even mouseup) by displaying a menu.

Source/WebKit/mac:

* WebCoreSupport/WebFrameLoaderClient.mm: (NetscapePluginWidget::handleEvent): Return true
for contextmenu event, so that plug-ins won't get a default WebKit context menu. We can't
know if the plug-in is handling mousedown (or even mouseup) by displaying a menu.

Source/WebKit2:

* WebProcess/Plugins/Netscape/NetscapePlugin.cpp: (WebKit::NetscapePlugin::handleContextMenuEvent):
* WebProcess/Plugins/Netscape/NetscapePlugin.h:
* WebProcess/Plugins/Plugin.h:
* WebProcess/Plugins/PluginProxy.cpp: (WebKit::PluginProxy::handleContextMenuEvent):
* WebProcess/Plugins/PluginProxy.h:
* WebProcess/Plugins/PluginView.cpp: (WebKit::PluginView::handleEvent):
Return true when handling contextmenu event, so that plug-ins won't get a default WebKit
context menu. We can't know if the plug-in is handling mousedown (or even mouseup) by
displaying a menu.

* WebProcess/Plugins/PDF/BuiltInPDFView.h:
* WebProcess/Plugins/PDF/BuiltInPDFView.cpp: (WebKit::BuiltInPDFView::handleContextMenuEvent):
PDF is the only "plug-in" that wants default WebKit menu now.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@98027 268f45cc-cd09-0410-ab3c-d52691b4dbfc