Add finalizer to JSObject https://bugs.webkit.org/show_bug.cgi?id=70336
Reviewed by Darin Adler.

* heap/MarkedBlock.cpp:
(JSC::MarkedBlock::callDestructor): Skip the call to the destructor 
if we're a JSFinalObject, since the finalizer takes care of things.
* runtime/JSCell.h:
(JSC::JSCell::~JSCell): Remove the GC validation due to a conflict with 
future changes and the fact that we no longer always call the destructor, making 
the information provided less useful.
* runtime/JSObject.cpp:
(JSC::JSObject::finalize): Add finalizer for JSObject.
(JSC::JSObject::allocatePropertyStorage): The first time we need to allocate out-of-line
property storage, we add a finalizer to ourself.
* runtime/JSObject.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@98123 268f45cc-cd09-0410-ab3c-d52691b4dbfc