Caret is drawn in the wrong place in multi-column blocks https://bugs.webkit.org/show_bug.cgi?id=70662
Reviewed by Gavin Barraclough.

* manual-tests/caret-in-columns.html: Added.
* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::paintCaret): Removed the call to offsetForContents(). That function takes
a point in local coordinates, whereas this function was applying it to a point in painting root
coordinates. The desired effect was only to undo the scroll adjustment done by the caller,
paintObject().
(WebCore::RenderBlock::paintObject): Pass the original, rather than scroll-adjusted, paint offset
to paintCaret().



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@98191 268f45cc-cd09-0410-ab3c-d52691b4dbfc