compositing/video tests time out on Lion https://bugs.webkit.org/show_bug.cgi?id=70448
Reviewed by Simon Fraser.

Source/WebCore:

AVFoundation will not begin loading or decoding media until a consumer for that media is present.
For video-only media, this means that an AVPlayerLayer must be created and attached to the
AVPlayer before the load state will ever reach "canPlayThrough". Once the metadata for a media
is loaded, create a AVPlayerLayer if the media has a video track.

* platform/graphics/avfoundation/MediaPlayerPrivateAVFoundation.cpp:
(WebCore::MediaPlayerPrivateAVFoundation::updateStates):

LayoutTests:

* compositing/video/video-poster.html: Seek, instead of play, to trigger
    creation of a layer
* http/tests/media/video-load-twice.html: Use waitForEventOnce so that
    canPlayThrough listener is triggered only once.
* media/video-test.js: Add a new function, waitForEventOnce.
(waitForEventOnce):
(waitForEvent._eventCallback):
(waitForEvent):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@98359 268f45cc-cd09-0410-ab3c-d52691b4dbfc