SVG Filter on a group doesn't invalidate when children are moved https://bugs.webkit.org/show_bug.cgi?id=70044 <rdar://problem/10281530>
Reviewed by Nikolas Zimmermann.

Call SVGResourcesCache::clientLayoutChanged whenever the element or its children need layout. Previously,
invalidation was only performed if the element itself needed layout; now we also invalidate if any child
needs layout and there is a filter applied, as the cached filter result can depend on the layout of children.

Test: svg/filters/invalidate-on-child-layout.svg

* rendering/svg/RenderSVGContainer.cpp:
(WebCore::RenderSVGContainer::layout):
* rendering/svg/SVGResourcesCache.cpp:
(WebCore::SVGResourcesCache::clientLayoutChanged):

Add a test that ensures that changing the layout of a child of a filtered element invalidates the filter.

* platform/mac/svg/filters/invalidate-on-child-layout-expected.png: Added.
* platform/mac/svg/filters/invalidate-on-child-layout-expected.txt: Added.
* svg/filters/invalidate-on-child-layout.svg: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@98989 268f45cc-cd09-0410-ab3c-d52691b4dbfc