DFG 32_64 - registers cannot be reused arbitrarily if speculation failures are possible https://bugs.webkit.org/show_bug.cgi?id=71684
Reviewed by Filip Pizlo.

Currently in DFG JIT, we try to reuse the physical register of an
operand for temporary usage if the current use of the operand is the
last use. But sometimes this can be wrong, for example if there are
possible speculation failures and we need to fallback to baseline JIT,
the value of the operand which is supposed to be hold in the physical
register can be modified by register reusing. The fixes the last
inspector failures in layout test on Mac 32-bit if switching on DFG.

* dfg/DFGSpeculativeJIT32_64.cpp:
(JSC::DFG::SpeculativeJIT::emitObjectOrOtherBranch):
(JSC::DFG::SpeculativeJIT::compile):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99476 268f45cc-cd09-0410-ab3c-d52691b4dbfc