[Qt] Change semantics of the haveQt() function to match API promises
Having Qt 5 does not imply that we have Qt 4, from an API point of
view, as a major version can in theory break BC/SC. Our minimum
version of Qt 4 will always be the latest released version of Qt 4,
so the only place we should use haveQt(4, x) is when checking for
the minimum requirements.

Reviewed-by Simon Hausmann.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99843 268f45cc-cd09-0410-ab3c-d52691b4dbfc