[Qt] Add a way to have experimental features in WebKit2 https://bugs.webkit.org/show_bug.cgi?id=67707
Reviewed by Kenneth Rohde Christiansen.

.: 

Add the new private QML module in the build.

* Source/QtWebKit.pro:

Source/WebKit/qt: 

This patch is adding a new module for accessing private
method and property from the view. It requires an import
QtWebKit.private in QML to be used.

* declarative/private/plugin.cpp: Added.
(WebKitQmlPluginPrivate::registerTypes):
* declarative/private/private.pro: Added.
* declarative/private/qmldir: Added.

Source/WebKit2: 

Add new QML extensions for our public view. This extension
returns the private object of the view so that the user could
use unstable/unreleased APIs. This requires our private class
to be exported and QObject subclass.

* Target.pri:
* UIProcess/API/qt/qquickwebview.cpp:
* UIProcess/API/qt/qquickwebview_p.h:
* UIProcess/API/qt/qquickwebviewprivateextension.cpp: Added.
(QQuickWebViewPrivateExtension::QQuickWebViewPrivateExtension):
(QQuickWebViewPrivateExtension::viewPrivate):
* UIProcess/API/qt/qquickwebviewprivateextension_p.h: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@99952 268f45cc-cd09-0410-ab3c-d52691b4dbfc