[chromium-mac] Enable vertical text using Skia https://bugs.webkit.org/show_bug.cgi?id=72137
Source/WebCore:

Use Skia to draw vertical text. This is much
faster and has higher fidelity than the old method
of drawing text on a path.

The graphics context passed to Skia has been
rotated 90 degrees but the character advances
have not, so it is necessary to unrotate the canvas,
and re-rotate the positions.

This generates correct output (or, at least,
consistent with Chromium CG on Mac) for all vertical
text tests, one of which is mentioned below.

Reviewed by Stephen White.

Test: fast/writing-mode/text-orientation-basic.html

* platform/graphics/skia/FontSkia.cpp:
(WebCore::setupPaint):
(WebCore::Font::drawGlyphs):

LayoutTests:

Ignore vertical text tests on Skia on Mac for now.

Reviewed by Stephen White.

* platform/chromium/test_expectations.txt:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@100283 268f45cc-cd09-0410-ab3c-d52691b4dbfc