repurpose the flexbox bot for grid layout https://bugs.webkit.org/show_bug.cgi?id=72557
Reviewed by Ryosuke Niwa.

Since new flexbox is enabled everywhere, we don't need this bot
anymore.  However, work on grid layout has begun, so we can use this
bot to compile with ENABLE_CSS_GRID_LAYOUT.

* BuildSlaveSupport/build.webkit.org-config/config.json: Just rename
the bot on the waterfall, I will change the slave config to enable
the compile flag and tests.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@100548 268f45cc-cd09-0410-ab3c-d52691b4dbfc