[EFL] Clean up the use of DATA_DIR in the buildsystem https://bugs.webkit.org/show_bug.cgi?id=72681
Patch by Raphael Kubo da Costa <kubo@profusion.mobi> on 2011-11-17
Reviewed by Daniel Bates.

.:

* Source/cmake/OptionsEfl.cmake: Rename DATA_DIR to DATA_INSTALL_DIR
and add it to the cache; add a variable with the path of the generated
theme and remove the -DDATA_DIR definition, it was moved to
PlatformEFL.cmake in WebCore.

Source/WebCore:

Add the -DDATA_DIR definition here instead of defining it globally in
OptionsEfl.cmake, as WebCore is the only place which needs it.

No new tests, this is a buildsystem change.

* PlatformEfl.cmake:

Source/WebKit/efl:

Instead of messing with BUILD_DATA_DIR and PARENT_SCOPE, just use
THEME_BINARY_DIR as defined in OptionsEfl.cmake.

* CMakeListsEfl.txt:

Tools:

Instead of adding -DDATA_DIR in all scopes, only define it in
WebCore/PlatformEFL.cmake, as WebCore is the only place that uses
it.

We can then define DATA_DIR to the generated theme directory for
EWebLauncher, since it's what we want here.

* CMakeListsEfl.txt:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@100722 268f45cc-cd09-0410-ab3c-d52691b4dbfc