[GTK] Accessibility API tests not loading Gtk's accessibility implementation https://bugs.webkit.org/show_bug.cgi?id=72732
Reviewed by Philippe Normand.

Ensure that the GAIL module is being loaded by avoiding using
gtk_test_init() directly, but a modified version of it.

* tests/testatk.c:
(testWebkitAtkParentForRootObject): Use a normal GtkWidget (a
GtkWindow) as the parent container for the webView, instead of
using a dummy webView for that, which was hackish.
(testWebkitAtkSetParentForObject): Use normal GtkWidgets instead
of dummy webViews (a GtkWindow and a GtkButton).
(initializeTestingFramework): New function, identical to
gtk_test_init(), but loading GAIL as the only GTK module.
(main): Use initializeTestingFramework().

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@100783 268f45cc-cd09-0410-ab3c-d52691b4dbfc