Showing the data overlay in OpenStreetMap doesn't work, zooming partially broken https://bugs.webkit.org/show_bug.cgi?id=71505
Reviewed by Gavin Barraclough.
        
It turns out that we were corrupting phi nodes in case of overflow. The bug is
really obvious, but producing a test case that causes the badness is hard. Even
when the phi nodes do get corrupt, there's more that has to happen before it
causes incorrect execution - and I wasn't able to reproduce in any kind of
sensible reduced case.

* dfg/DFGByteCodeParser.cpp:
(JSC::DFG::ByteCodeParser::processPhiStack):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@100972 268f45cc-cd09-0410-ab3c-d52691b4dbfc