ews bots should pass --force to update-webkit-chromium https://bugs.webkit.org/show_bug.cgi?id=73230
Reviewed by Adam Barth.

This makes it less likely for gclient changes to break the bots (e.g.,
if a DEPS repository moves or is switched to a branch).

Also pass --force when using |build-webkit --update-chromium| since this
appears to only be used by the bots.

* Scripts/update-webkit:
* Scripts/webkitdirs.pm:
(determineIsChromium): Add --force-update to update-webkit --chromium.
(forceChromiumUpdate):
(buildChromium): Pass --force to update-webkit-chromium.
* Scripts/webkitpy/tool/steps/update.py:
(Update.run): Add --force-update if non-interactive (i.e., bots).
* Scripts/webkitpy/tool/steps/update_unittest.py:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101279 268f45cc-cd09-0410-ab3c-d52691b4dbfc