https://bugs.webkit.org/show_bug.cgi?id=72551 When the recommended scrollbar style changes, WKView's tracking options should adjust accordingly -and corresponding- <rdar://problem/10409328>
Reviewed by Darin Adler.

Source/WebCore: 

This new ChromeClient function is called when the recommended scrollbar style 
changes. This way, WebKit can respond to the change by adjusting its mouse 
tracking.
* page/ChromeClient.h:
(WebCore::ChromeClient::recommendedScrollbarStyleDidChange):

Existing ScrollableArea function scrollbarStyleChanged() now takes an int 
indicating the new scrollbar style and a bool indicating whether it is necessary 
to force an update. It used to be the case that this function was ONLY used to 
force an update (and only called when an updated was needed), but now that it must 
also call into the ChromeClient, it is necessary to include a bool tracking 
whether we need to force an update. New implementation on FrameView is responsible 
for calling ChromeClient, and then that calls into the pre-existing ScrollView 
function for the forceUpdate part.
* page/FrameView.cpp:
(WebCore::FrameView::scrollbarStyleChanged):
* page/FrameView.h:
* platform/ScrollView.cpp:
(WebCore::ScrollView:: scrollbarStyleChanged):
* platform/ScrollView.h:
* platform/ScrollableArea.h:
(WebCore::ScrollableArea::scrollbarStyleChanged):
* platform/mac/ScrollAnimatorMac.mm:
(WebCore::ScrollAnimatorMac::updateScrollerStyle):

Source/WebKit2: 

These new functions take care of passing along the 
recommendedScrollbarStyleDidChange() message that originates in the ChromeClient. 
* UIProcess/API/mac/PageClientImpl.h:
* UIProcess/PageClient.h:
* UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::recommendedScrollbarStyleDidChange):
* UIProcess/WebPageProxy.h:
* UIProcess/WebPageProxy.messages.in:
* WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::recommendedScrollbarStyleDidChange):
* WebProcess/WebCoreSupport/WebChromeClient.h:

This is where we actually respond to the recommendedScrollbarStyleDidChange 
message. We remove the existing tracking area and create a new tracking area with 
the appropriate tracking options.
* UIProcess/API/mac/PageClientImpl.mm:
(WebKit::PageClientImpl::recommendedScrollbarStyleDidChange):

BuiltInPDFView inherits from WebCore::ScrollableArea, so scrollbarStyleChanged() 
must now take two parameters like the one in ScrollableArea.
* WebProcess/Plugins/PDF/BuiltInPDFView.cpp:
(WebKit::BuiltInPDFView::scrollbarStyleChanged):
* WebProcess/Plugins/PDF/BuiltInPDFView.h:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101290 268f45cc-cd09-0410-ab3c-d52691b4dbfc