[Qt] QQuickWebView gets wrong position after reload https://bugs.webkit.org/show_bug.cgi?id=73292
Reviewed by Simon Hausmann.

The ensureContentWithinViewportBoundary, animates the current viewport
item into boundaries. That ofcourse breaks when we try to animate it,
given the initial-scale which hasn't been applied; so apply it.

Also put the QScroller settings code in the ctor as it never changes.

* UIProcess/qt/QtViewportInteractionEngine.cpp:
(WebKit::QtViewportInteractionEngine::QtViewportInteractionEngine):
(WebKit::QtViewportInteractionEngine::ensureContentWithinViewportBoundary):
(WebKit::QtViewportInteractionEngine::reset):
(WebKit::QtViewportInteractionEngine::applyConstraints):
* UIProcess/qt/QtViewportInteractionEngine.h:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101339 268f45cc-cd09-0410-ab3c-d52691b4dbfc