Original page URL is not set in quarantine information when downloading using context menu Save Linked File https://bugs.webkit.org/show_bug.cgi?id=73475 <rdar://problem/10500337>
Reviewed by Dan Bernstein.

Source/WebCore:

* WebCore.exp.in: Exported FrameLoader::setOriginalURLForDownloadRequest.

Source/WebKit2:

* WebProcess/WebProcess.cpp: (WebKit::WebProcess::downloadRequest): Set main document URL
for the request, so that CFNetwork would know where te file was downloaded. Note that this
is an easier case than downloading due to clicks, since we don't really need to build a host
only URL, and could use full page URL. However, the same function is used for consistency.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101607 268f45cc-cd09-0410-ab3c-d52691b4dbfc