Change HashMap implementation to use the pass type and peek type from traits for the mapped value https://bugs.webkit.org/show_bug.cgi?id=72474
Reviewed by Anders Carlsson.

* wtf/HashMap.h: Added ReferenceTypeMaker struct template. Get PassInType, PassOutType,
and PeekType from the traits of the mapped value instead of hard-coding them here.
Changed inlineAdd to take a reference to the PassInType instead of the PassInType itself,
to accomodate a PassInType that can't be copied. Use the store, peek, and passOut
functions from the traits as well.

* wtf/HashTraits.h: Updated GenericHashTraits and HashTraits for OwnPtr to include
PassInType, PassOutType, PeekType, store, passOut, and peek. Before this, the file had
an earlier version that was just PassType, PeekType, pass, and peek. Also commented
the HashTraits for RefPtr to foreshadow some work we can do there.

* wtf/RefPtrHashMap.h: Same changes as HashMap.h.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@101942 268f45cc-cd09-0410-ab3c-d52691b4dbfc