Get rid of KURL::KURL(ParsedURLStringTag, const char*); https://bugs.webkit.org/show_bug.cgi?id=73792
Reviewed by Andreas Kling.

In all cases where the constructor is used, the constructor
taking a String is as effective because the string is valid and converted
to String for m_string.

This patch remove the constructor KURL::KURL(ParsedURLStringTag, const char*)
and change the call sites that were using that constructor to build
empty URLs.

* dom/Document.cpp:
(WebCore::Document::initSecurityContext):
* loader/FrameLoader.cpp:
(WebCore::FrameLoader::init):
* loader/archive/cf/LegacyWebArchive.cpp:
(WebCore::LegacyWebArchive::create):
* page/DOMWindow.cpp:
(WebCore::DOMWindow::createWindow):
* platform/KURL.cpp:
* platform/KURL.h:
* platform/KURLGoogle.cpp:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@102080 268f45cc-cd09-0410-ab3c-d52691b4dbfc