Don't crash in StyleAttributeMutationScope if the style declaration's element has been GCed https://bugs.webkit.org/show_bug.cgi?id=74321
Reviewed by Ryosuke Niwa.

Source/WebCore:

In r101101, Rafael Weinstein added code to CSSMutableStyleDeclaration.cpp
which depended on isInlineStyleDeclaration returning true iff the
element it pointed to was non-null (it will be nulled-out if the
element is garbage collected).

Then, in r101172, Andreas Kling changed the semantics so that
isInlineStyleDeclaration only described the type of the declaration,
not the state of the related element.

This change updates Rafael's code with an explicit check that the
element is still alive.

Test: fast/dom/css-inline-style-declaration-crash.html

* css/CSSMutableStyleDeclaration.cpp:

LayoutTests:

* fast/dom/css-inline-style-declaration-crash-expected.txt: Added.
* fast/dom/css-inline-style-declaration-crash.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@102639 268f45cc-cd09-0410-ab3c-d52691b4dbfc