Audio file in video element has a size of 0x0 . https://bugs.webkit.org/show_bug.cgi?id=74738
Reviewed by Kenneth Rohde Christiansen.

Source/WebCore:

When the source of a video element has audio only, the intrinsic size of the video should
not be 0x0. Instead, it should be the same as as no media was loaded.

No new tests. An existing test is covering this case and was modified to reflect this change.

* rendering/RenderVideo.cpp:
(WebCore::RenderVideo::calculateIntrinsicSize):

LayoutTests:

Changed the expected result to reflect this change.

* media/audio-only-video-intrinsic-size-expected.txt:
* media/audio-only-video-intrinsic-size.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@103098 268f45cc-cd09-0410-ab3c-d52691b4dbfc